requirements.txt
fastapi
uvicorn
opencv-python
pillow
numpy
python-multipart
tflite-runtime
